refactor: apply same logic to destroy
