+++ pkg/filesystem/manifest_test.go
@@ -1,0 +1,57 @@
+// Copyright 2024 The Okteto Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package filesystem
+
+import (
+	"path/filepath"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func Test_cleanManifestPath(t *testing.T) {
+	var tests = []struct {
+		name     string
+		manifest string
+		expected string
+	}{
+		{
+			name:     "empty manifest",
+			manifest: "",
+			expected: ".",
+		},
+		{
+			name:     "absolute path to manifest file",
+			manifest: "/path/to/service/okteto.yml",
+			expected: "okteto.yml",
+		},
+		{
+			name:     "relative path to manifest file",
+			manifest: "./service/okteto.yml",
+			expected: "okteto.yml",
+		},
+		{
+			name:     "manifest within .okteto",
+			manifest: "/path/to/service/.okteto/okteto.yml",
+			expected: filepath.Clean(".okteto/okteto.yml"),
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			result := CleanManifestPath(tt.manifest)
+			assert.Equal(t, tt.expected, result)
+		})
+	}
+}
